requirements.txt
gunicorn~=20.1
prometheus-client~=0.13
pyyaml~=6.0
requests~=2.27
